upgrade libuv to ab8c3b85c1

+++ deps/uv/test/test-get-memory.c
@@ -26,7 +26,7 @@ TEST_IMPL(get_memory) {
   uint64_t free_mem = uv_get_free_memory();
   uint64_t total_mem = uv_get_total_memory();
 
-  printf("free_mem=%llu, total_mem=%llu\n", free_mem, total_mem);
+  printf("free_mem=%zu, total_mem=%zu\n", (size_t)free_mem, (size_t)total_mem);
 
   ASSERT(free_mem > 0);
   ASSERT(total_mem > 0);
